Move API builder scripts to a separate directory
